requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio @ file:///tmp/build/80754af9/anyio_1644481695334/work/dist
appdirs==1.4.4
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644553347904/work
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
Babel @ file:///croot/babel_1671781930836/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
billiard==3.6.4.0
binaryornot==0.4.4
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
boto3==1.24.59
botocore==1.27.59
brotlipy==0.7.0
cachetools==5.2.0
celery==5.2.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click @ file:///tmp/build/80754af9/click_1646056706450/work
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.5
cookiecutter==2.1.1
coverage==7.0.4
cryptography==39.0.0
cycler==0.11.0
debugpy @ file:///home/builder/ci_310/debugpy_1640789504635/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dictdiffer==0.9.0
diskcache==5.4.0
distlib==0.3.6
dpath==2.1.3
dulwich==0.20.50
dvc==2.38.1
dvc-data==0.28.4
dvc-gdrive==2.19.1
dvc-http==2.27.2
dvc-objects==0.14.0
dvc-render==0.0.15
dvc-s3==2.21.0
dvc-studio-client==0.3.0
dvc-task==0.1.8
dvclive==1.3.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
exceptiongroup==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
filelock==3.8.0
flake8==6.0.0
flatten-dict==0.4.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
flufl.lock==7.1.1
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future @ file:///home/builder/ci_310/future_1640790123501/work
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
grandalf==0.6
hydra-core==1.3.1
idna @ file:///croot/idna_1666125576474/work
iniconfig==2.0.0
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///croot/ipython_1670919316550/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
iterative-telemetry==0.0.6
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.2.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///opt/conda/conda-bld/jsonschema_1663375472438/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///croot/jupyter_console_1671541909316/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client @ file:///croot/jupyter_client_1671703053786/work
jupyter_core @ file:///croot/jupyter_core_1672332224593/work
jupyterlab @ file:///croot/jupyterlab_1672132689850/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server @ file:///croot/jupyterlab_server_1672127357747/work
kiwisolver==1.4.4
kombu==5.2.4
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1670466153205/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mccabe==0.7.0
mistune @ file:///home/builder/ci_310/mistune_1640791641754/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
multidict==6.0.4
munkres==1.1.4
mypy==0.991
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic @ file:///croot/nbclassic_1668174957779/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650290238894/work
nbconvert @ file:///croot/nbconvert_1668450669124/work
nbformat @ file:///croot/nbformat_1670352325207/work
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==2.8.8
notebook @ file:///croot/notebook_1668179881751/work
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
omegaconf==2.3.0
packaging==21.3
pandas==1.5.2
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.9.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
PyDrive2==1.15.0
pyflakes==3.0.1
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyOpenSSL==23.0.0
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyrsistent @ file:///home/builder/ci_310/pyrsistent_1640807196327/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==7.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-slugify==7.0.0
pytz @ file:///croot/pytz_1671697431263/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1662018252641/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rich==13.0.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2022.11.0
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
scmrepo==0.1.4
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
shortuuid==1.0.11
shtab==1.5.8
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio @ file:///home/builder/ci_310/sniffio_1640794799774/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.9.0
terminado @ file:///croot/terminado_1671751832461/work
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2 @ file:///croot/tinycss2_1668168815555/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
torchtext==0.14.1
torchvision @ file:///croot/torchvision_1670313549871/work
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1670526988650/work
vine==5.0.0
virtualenv==20.16.5
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client @ file:///home/builder/ci_310/websocket-client_1640795866898/work
wget==3.2
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354667/work
wrapt==1.14.1
yarl==1.8.2
zc.lockfile==2.0